Refactor and move neural network logic to a new module
Extracted the `nauthilus_call_neural_network` function from `geoip.lua` and moved it to a new `neural.lua` module under `features`. This improves code modularity and separation of concerns, making the GeoIP filter and neural network logic independently maintainable.
